Move http specific types to http.go
